pyproject.toml
[project]
name = "book-knowledge-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.83.0",
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.3",
    "langchain-text-splitters>=1.1.1",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "openai>=2.21.0",
    "plotly>=6.5.2",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
    "umap-learn>=0.5.11",
]
